Added beginnings of Status Parsing.
The model for the logic of the Status Parser is that of a DFA.  This means
it had to be heavily commented and the order of the if statements is
critical.  Unless I have an objection raised against this concept, I will
probably model the other parsers after this one.

I also made minor fixes (typos etc.) to the Status Effects data file.