Finished a working version of statusEffect parser.
The following is a partial list of features not currently working with the
parser: Immunities, about 6 or 7 abilities' statuses, categories (not
included yet but infrastructure is done).

This commit fixes several typos in the Status_Effects_Data.txt file.  This
commit also added a few properties (getters/setters) to the person and
playerCharacter classes.

The internalStatus and status classes are using a strange incomplete
constructor pattern that I dislike, but seems to work for now.  A code
review would help that as well as some scoping issues in the parser.