Added the Combat class.
The Combat class is inside the combat module.  It contains, or will
contain, many functions specific to battles and will be dense with game
rule logic.  It has only a few of the methods implemented at the moment.

I also made small changes to other files in response to thoughts of
details the combat class will need, or will want to rely on.