Updated the way statuses are instantiated.
The internalstatus.py file has a block of comments at the top that
explains the proper use and instantiation of Statuses and
InternalStatuses.  At this point, the creation and population of needed
fields has been solidified with the possible exception of the element
property being conveyed to the InternalStatuses.

Additionally, the format for the Status Effects Data file has been altered
slightly to include a RECURRING field that is typically False, but when
True lets the InternalStatus know that it needs to apply its action during
every upkeep.