Added get/setters to many vars in playercharacter
A very modest change was made in combat.py as well.

A large number of getters and setters remain to be added to the
playercharacter module/class.  Many of the new getters/setters do not have
lengthy comments (if any at all).  I simply don't have the time to give
detailed information on every variable here; at least not now.