Add a traps parser that is functional.
In addition to the functional traps parser, The trap class should be fully
functional for hostile traps (untested).

The dice and combat classes had some unnoticed syntax errors but they are
now fixed.

The statuseffectsparser was using a hard-coded file to load from; that has
been fixed although the main method there used for testing will use a
hard-coded value.