Moved most of PC to person.py
In order to clarify the heirarchy for Entity objects I needed to shift
most of playercharacter into person to ensure I can inherit from those
fields/methods with Monster which I have yet to write.

I also changed a couple of methods in weaponsparser to be a little cleaner
by using the import x as y thingi.