Added __init__ to person.py
The constructor for Person objects requires a dictionary given from a
parser.  It accepts any number of parameters this way in any order.
Certain values, such as 'startingHP' must be supplied, or an Error will be
thrown.

I swapped the logic in the Person class to accommodate for the traditional
Y-axis direction used in computer science.  It only affected two methods.

The logic used to determine the 'total' of a derived statistic had to be
changed.  Previously, the total was adding together the three basic
factors (equipment, base, and status) and was done.  However, I realized
due to the dependencies between the derived statistics and the primary
statistics that I needed to calculate the totals based on the primary
statistics whenever they are viewed and reserve the 'base' value for
racial bonuses, traits and the like.

The next step will be to make similar changes to the playercharacter
class, including its own __init__ function.