Added Entity and some levelup stuff.
The Entity class is mostly just a shell at the moment.  It is the class
that most interactable objects will need to inherit from.

Some levelup mechanics have been added in accordance with the new wiki
entry on Levelup.  The UI elements will need to be integrated with it to
continue.