Added equippeditems module.
This new module acts as a container for all the Equipment (Armor/Weapons)
currently in use by a player.  It will handle equipping new gear and will
help with the automatic adjusting of stats from doing so.

Minor changes had to be made to the playercharacter, person, and combat
modules to accommodate this new module's role.

The module is not finished, but has most of its work done as it is a
pretty simple class.