Began infrastructure for listeners.
This will require a lot of work, and currently the 'backstab' parameter
still needs to be added to several abilities in the Spell_Templates.txt
file.

Importantly, Dual weapon attacks are now *almost* supported by using a
wrapper for weaponAttack that detects if the player is using two weapons
or one.