Began work on Traits.
My current work is in conflict with the PassiveAbility class, but that can
be fixed without too much work.  I am switching from using single
'onStrings' to lists of 'onStrings'.  (offStrings as well.)

I created a helper method that didn't exist but was being called from the
person module; the helper method resides in the equippeditems.py file and
sends back a string indicating the string representation of the armor
grade being worn by a player.

Also, we will need to keep track of how many attacks (and later movements)
each player took during the current turn, and the previous turn.  I am
planning on just keeping track of those things in a two piece array that
gets updated whenever an attack (or move) is performed by that character,
and will be updated during the 'upkeep' phase right before a player's
turn.