Added missing Fighter traits.
I am still missing two 'terrain-based' functions that I will need to fully
implement the Fighter traits correctly.

Also, I changed the force to be expecting ints instead of floats.  There
may be some code that still expects the base to be 1.0 not 100. We will
just have to be aware of this and see if bugs crop up.