Began Spell class/module.
The spell class will contain easy-to-call methods to check if a spell is
capable of being cast (due to resources, range, cooldowns, etc.) as well
as providing simple wrappers to actually call the spell.  It currently
relies on some simple distance measuring that I have not implemented, and
that probably belongs in another developer's domain more than mine.

I added a healing function to combat.py as well as a minor fix to the
lowerHP function.  While we could make direct calls to modifyResource,
these wrapper calls will help with the listener pattern.  It will lead
to code that is both more readable as well as slightly faster, I believe.

The Spell_Templates.txt file will is not going to be used (directly) to
perform our spell/ability logic.  Rather, it will be coded in a somewhat
easy-to-mod format in the spell class (for the spells) and in a currently
undeveloped class (for the abilities).  I would like to keep it in this
branch for now as I am using it as a reference.