Added categories to statuses.
This change only affected the data files.

I rolled back one change in person.py that I realized would introduce a
bug.  The movementAPOverride needs to be able to function without a
duration and without a number of moves in the event that a character is
stealthed because it is limited by neither in that case.