Implemented important listeners.
This involved small changes to many files and large changes to
broadcast.py, and listener.py, and combat.py.  passiveability.py switched
such that its actions are no longer static methods.  This requires passing
the ability itself to the listeners etc. which is a pain, but I see no
other way to do it without a ton of rework.