Added inventory.py .
The Inventory class is very simple at the moment, it mostly just contains
a list that should contain all items not currently equipped on the player.
We will need an InventoryItem class; perhaps it will belong in the same
module.

I made a change to some other classes, primarily to ensure that items that
are unequipped are placed in the player's inventory, instead of being
lost.