Added magical properties.
The magical properties class contains a lot of logic, and none of it has
been tested at all.  There is also a small performance worry that will
have to be checked on.

Currently the magical property class does not contain on-hit properties
and one other.  Also, the functions to generate magical properties do not
yet support treasure chests that give gear based on the class of each
character.