Finished most of onHitEffects back end.
I still need to add the on hit effects to the magical properties module to
allow them to be selected on item generation.  Also, I need a method in
the playercharacter module to remove an on hit effect without having a
reference to the actual effect.  Finally, I need to add a tiny bit of
functionality to the combat module to get it to apply every on hit effect
from the attacking player when an attack lands.