Fully implemented magical equipment.
The playercharacter module was updated to fully support magical equipment.
I also made what progress I could on the levelup process.  Fixed a bug in
the equippeditems module where functions were returning an object when I
needed them to return a list of objects (even if there is only one.)