Added Spellsword up to level 2.
This involved creating a new listener and more importantly a new
broadcaster.  This broadcast sends out a shout every time AP, MP, or HP is
altered.  This may be a minor performance issue later.  Currently, no
abilities care about AP levels, so a simple optimization would be to
remove the AP shout at least.  But for now, I'll keep it in case we care
about AP levels later.

The only abilities I can think of that currently listen for those levels
to change in that fashion are temporary ones whose listeners are not
always present on a character (only for the duration of a status or
whatnot.)  That helps reduces the overhead somewhat.

Also, the ability, 'Martial Mode' required a lot of plumbing; it was the
first ability to require setting up a listener on an active ability.
Check out the code for it in the ability.py file by searching for
_martialMode as well as _martialModeDisable .