Centralized the inRange function.
The checks for inRange in ability and spell.py have been moved to
Person.py.  Also, a range of -1 indicates to use the default weapon range
(which may be modified by the Marksman bonus.)  Monsters will need to have
an attackRange property added to them to make the functions work for them
as well.