Fixed more float(), added img to monster
The monster does not have an image but the monstersparser will now require
that tag.  The field will need to be modified with an actual filename in
the monster_data file at a later time.

The combat class has been purged of the float division bugs, as has the
person module.