+most work on lvl1 Tactician.
Missing the functionality to actually lay down the trap on the combat
arena.  Changed Sticky Trap (Ranger) to have 2 charges for balance
reasons.  Changed the location of where charges are determined from the
abilities themselves to the trap dict in trap.py .

Added a method to check whether a trap should be triggered when a Person
steps on it that does some comparisons based on their team and whether the
trap is actually a Tactician's Favor.  I think we'll want Favors to
trigger whenever passed over, not just when a player ends their turn on
one.