Monsters now attack you.
Many, many bugs were discovered in my code.  The biggest thing is that A)
we have monsters that will attack and use abilities on you if you move
into their range, and B) there are still weird "dead-zones" that will
screw with the monster AI as well as the movement AP cost.  That needs to
be fixed!