Added some status-backend.
We don't actually have images for status effects in this repository.
However, the files referenced do exist but are .svg files at the moment.
They will need to be converted by hand.  Eventually some of them will need
to be recolored.  It remains to be seen if some of them need to be resized
or if that can be done dynamically.