Fixed bugs, lowered difficulty.
Monster HP is down abou 10% (for level 1 monsters.)  Monster basic attack
damage is down about 20% (for level 1 monsters) and other stats are down
slightly for monsters up to level 3.

Scrolling wouldn't have worked the way I had it set up, because it was
foolishly trying to do it on the server side.  Also, some code that didn't
do anything but could be a source for confusion was removed.