Treasure chests on server only
Server now sends a command to clients to add a treasure chest.
The chest stays on the screen until you leave the pane.  When you
come back it is gone.  We can change this to fade it out, but
that isn't priority.
New players entering a pane will not see open chests.