Fixed edge tiles bug
After "fixing" the phantom objects in the combat pane bug, I
introduced this one (or rather uncovered it). Basically when I
would load the surrounding panes I made sure not to store the
entities into the tiles or load images (saved on processing).
However, any call to the tile asking if it is passable would
return True because of this.  So, when I merge tiles, I add
the keys to the entities on the edges to the tile of the pane I
am merging to and changed the is_passable() function of tile to
check this new list (entity_keys) against a list of known obstacles.