A buggy portal implementation
Kyle, I've modified your server code, see lines 71-76
and you'll find the parts that will need some serious updating.
The basic idea is there, it's super buggy and you'll see
that I'm not handling things right.  So if you get a chance, take
a look and feel free to mess around with it.

See buildings.py line 105 to mess with the actual portal entities
that I add to the doorways of the houses.  I've hard coded in
the center of pane (0, 0) as the location of the portal.  If you go
to (0, 2) there are some houses there to play with.  Pane switching
isn't handled properly either.