Bug fixes, +abil, balance.
Changed the way the poison hit method works.  The implementation details
are not interesting, but it now displays darkgreen text for
player-->monster poison rolls and red for monster-->player poison rolls.

Druid's level 2 ability is functioning now and was balanced to deal about
8 less damage per turn.

Thief trait "Clever" was underpowered, a tiny meleeDodge bonus was added
to it.

Ranger trait "Nature Training" was underpowered, a small poison rating
bonus was added to it, making it particularly appealing for Assassins and
Druids.