Fixed monster bug with save/load
Monsters were being lost when they wandered onto a town pane when
the game was saved.  This fix will load monsters on town panes.

We may not want this behaviour... but we may not want monsters
wandering onto towns in the first place.  So if we don't want this
we should fix that first.