+Anarchist stuff
The missing level 5 ability Grab Bag has been added but was altered to be
made much more powerful in most situations.  This has been reflected on
the wiki.

A new passive ability giving the Anarchist a passive +20% trap damage has
been added in both wiki and code to make it a more balanced class.  Due to
his high AP costs with both ranged and melee attacks, no spells, and only
medium armor, he needs to be THE trap character for trap enthusiasts.