Shop caching and mvd to serverside.
It would be simple to save shop information on the world save now.
However, the shops are independent of the world seed provided, so that
might be a little odd.  That could be changed though.

Changed enter to be a valid accept key for the save menu, and
ability/spell selection in combat.