Fix lag
trapList was originally generating a region in pane (0, 0), a.k.a., the combat arena.  I made a change so that I could use this method to show paths on the world map with a keybinding, and changed one corner of the generated region to be in the pane the player is in.  However, the other corner was still rooted firmly in pane (0, 0).  This fix moves the other corner of this region to the pane the player is in, and should virtually eliminate this long standing bug.  It's still a little inefficient, and I'm going to further optimize it tonight.  But this should still take care of the vast majority of the problem.